Refactor Value implementation.
