+++ minijinja/src/functions.rs
@@ -228,7 +228,7 @@ mod builtins {
     use std::collections::BTreeMap;
 
     use crate::error::ErrorKind;
-    use crate::value::{MapType, ValueRepr};
+    use crate::value::{MapType, ValueBuf};
 
     /// Returns a range.
     ///
@@ -291,8 +291,8 @@ mod builtins {
     #[cfg_attr(docsrs, doc(cfg(feature = "builtins")))]
     pub fn dict(value: Value) -> Result<Value, Error> {
         match value.0 {
-            ValueRepr::Undefined => Ok(Value::from(BTreeMap::<bool, Value>::new())),
-            ValueRepr::Map(map, _) => Ok(Value(ValueRepr::Map(map, MapType::Normal))),
+            ValueBuf::Undefined => Ok(Value::from(BTreeMap::<bool, Value>::new())),
+            ValueBuf::Map(map, _) => Ok(Value(ValueBuf::Map(map, MapType::Normal))),
             _ => Err(Error::from(ErrorKind::InvalidOperation)),
         }
     }
